# feat: optimize secret referencing performance