--- a/APKBUILD
+++ b/APKBUILD
@@ -1,6 +1,6 @@
 # Maintainer: Phase <info@phase.dev>
 pkgname=phase
-pkgver=1.19.4
+pkgver=1.19.5
 pkgrel=0
 pkgdesc="Phase CLI"
 url="https://phase.dev"