--- a/phase_cli/utils/const.py
+++ b/phase_cli/utils/const.py
@@ -1,7 +1,7 @@
 import os
 import re
 
-__version__ = "1.19.4"
+__version__ = "1.19.5"
 __ph_version__ = "v1"
 
 description = "Securely manage application secrets and environment variables with Phase."