… tests/test_secret_referencing_extended.py