This PR improves secret referencing resolution performance by using a client-side in-memory cache, fetching secrets in parallel and significantly reducing the number of API calls made to the Phase API. This dramatic reduction in API calls would mean fewer run-ins with API rate limits and further reduction of the number of secret audit logs being generated. Added recursive secret resolution support and extended partial env name matching logic. 

This is especially helpful for users with larger number of cross application referenced secrets (very often more than 100+ secrets) for use cases involving the user of another applications environment as a shared or common application.


Benchmarks: 

Example application and environment structure.

```tf
terraform {
  required_providers {
    phase = {
      source  = "phasehq/phase"
    }
  }
}

provider "phase" {
  phase_token = var.phase_access_token
}

variable "phase_access_token" {
  description = "Phase personal access token"
  type        = string
  sensitive   = true
  default     = null
}

variable "app_a_id" {
  type    = string
  default = "a6ae711a-7698-45d2-910e-af2f3dd7a05a"
}

variable "app_b_id" {
  type    = string
  default = "c660dd1c-9a75-4975-8a77-8df0603a5852"
}

variable "app_c_id" {
  type    = string
  default = "cfc93688-e6ae-4187-926e-7775bb1a2436"
}

variable "app_d_id" {
  type    = string
  default = "86ca4282-4ab7-494c-bd40-a27bdbf209a8"
}

variable "app_e_id" {
  type    = string
  default = "f664a249-2a9e-4fd8-97e0-d36b59f8e14f"
}

variable "environment" {
  type    = string
  default = "development"
}

locals {
  max_per_app = 500
  group_size  = 125
  app_letters = ["B", "C", "D", "E"]

  b_count = local.max_per_app
  c_count = local.max_per_app
  d_count = local.max_per_app
  e_count = local.max_per_app

  a_count = local.max_per_app
}

resource "phase_secret" "app_b" {
  count       = local.b_count
  app_id      = var.app_b_id
  env         = var.environment
  path        = "secret_ref"
  key         = "SECRET_${count.index + 1}"
  value       = "FROM APP B DEVELOPMENT"
}

resource "phase_secret" "app_c" {
  count       = local.c_count
  app_id      = var.app_c_id
  env         = var.environment
  path        = "secret_ref"
  key         = "SECRET_${count.index + 1}"
  value       = "FROM APP C DEVELOPMENT"
}

resource "phase_secret" "app_d" {
  count       = local.d_count
  app_id      = var.app_d_id
  env         = var.environment
  path        = "secret_ref"
  key         = "SECRET_${count.index + 1}"
  value       = "FROM APP D DEVELOPMENT"
}

resource "phase_secret" "app_e" {
  count       = local.e_count
  app_id      = var.app_e_id
  env         = var.environment
  path        = "secret_ref"
  key         = "SECRET_${count.index + 1}"
  value       = "FROM APP E DEVELOPMENT"
}

resource "phase_secret" "app_a" {
  count       = local.a_count
  app_id      = var.app_a_id
  env         = var.environment
  key         = "SECRET_${count.index + 1}"
  value       = format(
    "$${%s::%s.secret_ref/SECRET_%d}",
    local.app_letters[floor(count.index / local.group_size)],
    var.environment,
    (count.index % local.group_size) + 1,
  )
  depends_on  = [
    phase_secret.app_b,
    phase_secret.app_c,
    phase_secret.app_d,
    phase_secret.app_e,
  ]
}
```


Before:

```fish
...
SECRET_319="FROM APP D DEVELOPMENT"
SECRET_11="FROM APP B DEVELOPMENT"
SECRET_397="FROM APP E DEVELOPMENT"
SECRET_25="FROM APP B DEVELOPMENT"
SECRET_167="FROM APP C DEVELOPMENT"
phase secrets export  1.66s user 0.28s system 3% cpu 50.425 total
```

After:

```fish
...
SECRET_319="FROM APP D DEVELOPMENT"
SECRET_11="FROM APP B DEVELOPMENT"
SECRET_397="FROM APP E DEVELOPMENT"
SECRET_25="FROM APP B DEVELOPMENT"
SECRET_167="FROM APP C DEVELOPMENT"
phase secrets export  0.60s user 0.08s system 4% cpu 13.677 total
```